+++ lib/generators/amphtml/helpers_generator.rb
@@ -11,13 +11,16 @@ module Amphtml
 
             unless options[:split]
                 source_root File.expand_path("../../templates/helpers", __FILE__)
-                def create_views
+                def create_helpers
                     template "asset_tag_helper.rb", "app/helpers/asset_tag_helper.rb"
+
+                    template "amp/component_tag_helper.rb", "app/helpers/amp/component_tag_helper.rb"
+                    template "amp/social_tag_helper.rb", "app/helpers/amp/social_tag_helper.rb"
                 end
             else
                 source_root File.expand_path("../../templates/helpers/split", __FILE__)
-                def create_views
-                    template "asset_tag_helper.rb", "app/helpers/asset_tag_helper.rb"
+                def create_helpers
+                    # copy above
                 end
             end