+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * enhancements
     * allow custom CSS in views
     * warn when a `stylesheet_link_tag` or `javascript_include_tag` is being used
+    * improve `HelpersGenerator`
+* bug fixes
+    * improve `TagHelper`'s
 
 ### 0.1.5 (Pre-release) - 2017-03-13